backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
schedule==1.2.0
gitpython==3.1.37
python-dotenv==1.0.0
